requirements.txt
chroma-hnswlib==0.7.6
chromadb==0.6.3
langchain==0.3.23
langchain-chroma==0.2.2
langchain-community==0.3.21
langchain-core==0.3.51
langchain-huggingface==0.1.2
langchain-openai==0.3.12
langchain-text-splitters==0.3.8
langchainhub==0.1.21
openai==1.70.0
pandas==2.2.3
python-dotenv==1.1.0